+++ libs/ngx-editorjs-demo/src/lib/plugins/image/image.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { PluginImage } from './image.plugin';
-
-@NgModule({
-  providers: [PluginImage]
-})
-export class ImageModule {}