+++ libs/ngx-editorjs-demo/src/lib/plugins/link/link.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { PluginLink } from './link.plugin';
-
-@NgModule({
-  providers: [PluginLink]
-})
-export class LinkModule {}