+++ libs/ngx-editorjs-demo/src/lib/plugins/marker/marker.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { PluginMarker } from './marker.plugin';
-
-@NgModule({
-  providers: [PluginMarker]
-})
-export class MarkerModule {}