+++ libs/ngx-editorjs-plugins/src/index.ts
@@ -1,1 +1,17 @@
-export * from './lib/ngx-editorjs-plugins.module';
+export { NgxEditorjsPluginsModule } from './lib/ngx-editorjs-plugins.module';
+export { PluginCodeModule } from './lib/plugins/code/code.module';
+export { PluginCode } from './lib/plugins/code/code.plugin';
+export { PluginHeaderModule } from './lib/plugins/header/header.module';
+export { PluginHeader } from './lib/plugins/header/header.plugin';
+export { PluginImageModule } from './lib/plugins/image/image.module';
+export { PluginImage } from './lib/plugins/image/image.plugin';
+export { PluginLinkModule } from './lib/plugins/link/link.module';
+export { PluginLink } from './lib/plugins/link/link.plugin';
+export { PluginListModule } from './lib/plugins/list/list.module';
+export { PluginList } from './lib/plugins/list/list.plugin';
+export { PluginMarkerModule } from './lib/plugins/marker/marker.module';
+export { PluginMarker } from './lib/plugins/marker/marker.plugin';
+export { PluginParagraphModule } from './lib/plugins/paragraph/paragraph.module';
+export { PluginParagraph } from './lib/plugins/paragraph/paragraph.plugin';
+export { PluginSimpleImageModule } from './lib/plugins/simple-image/simple-image.module';
+export { PluginSimpleImage } from './lib/plugins/simple-image/simple-image.plugin';