+++ libs/ngx-editorjs-plugins/src/lib/plugins/code/code.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { PluginCode } from './code.plugin';
+
+/**
+ * A module that provides the default `EditorJS` `<code>` block tool.
+ * See [the GitHub repo](https://github.com/editor-js/code) for API details
+ */
+@NgModule({
+  providers: [PluginCode]
+})
+export class PluginCodeModule {}