+++ libs/ngx-editorjs-plugins/src/lib/plugins/code/code.plugin.ts
@@ -1,0 +1,17 @@
+import { Injectable } from '@angular/core';
+import Code from '@editorjs/code';
+import { ToolSettings } from '@editorjs/editorjs';
+import { BasePlugin } from '@tinynodes/ngx-editorjs';
+
+/**
+ * This plugin returns the tool settings for a `<code>` block tool for `EditorJS`
+ */
+@Injectable()
+export class PluginCode implements BasePlugin {
+  /**
+   * Returns the `ToolSettings` for this plugin
+   */
+  public plugin(): ToolSettings {
+    return Code;
+  }
+}