+++ libs/ngx-editorjs-plugins/src/lib/plugins/header/header.module.ts
@@ -2,9 +2,8 @@ import { NgModule } from '@angular/core';
 import { PluginHeader } from './header.plugin';
 
 /**
- * A module that provides the default `EditorJS` header plugin.
- * See [the GitHub repo](https://github.com/editor-js/header) for more details on
- * it's use
+ * A module that provides the default `EditorJS` header block tool.
+ * See [the GitHub repo](https://github.com/editor-js/header) for API details
  */
 @NgModule({
   providers: [PluginHeader]