+++ libs/ngx-editorjs-plugins/src/lib/plugins/list/list.module.ts
@@ -2,7 +2,8 @@ import { NgModule } from '@angular/core';
 import { PluginList } from './list.plugin';
 
 /**
- * Internal module to provide the List block editor
+ * A module that provides the default `EditorJS` `<li>` and `<ol>` block tool.
+ * See [the GitHub repo](https://github.com/editor-js/list) for API details
  */
 @NgModule({
   providers: [PluginList]