+++ libs/ngx-editorjs-plugins/src/lib/plugins/marker/marker.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { PluginMarker } from './marker.plugin';
+
+/**
+ * A module that provides the default `EditorJS` inline marker tool.
+ * See [the GitHub repo](https://github.com/editor-js/marker) for API details
+ */
+@NgModule({
+  providers: [PluginMarker]
+})
+export class PluginMarkerModule {}