+++ libs/ngx-editorjs-plugins/src/lib/plugins/paragraph/paragraph.plugin.ts
@@ -4,7 +4,7 @@ import { ToolSettings } from '@editorjs/editorjs';
 import { BasePlugin } from '../../types/plugins';
 
 /**
- * Plugin class to return the EditorJS plugin
+ * This plugin returns the tool settings for a `<p>` block tool for `EditorJS`
  */
 @Injectable()
 export class PluginParagraph implements BasePlugin {