+++ libs/ngx-editorjs/src/lib/services/editorjs.service.ts
@@ -9,10 +9,10 @@ import { PluginService } from './plugins.service';
 /**
  * The NgxEditorJSService provides control EditorJS instances via Angular.
  * The service has several maps:
- * * A map of `EditorJS` instances
- * * A map of `BehaviorSubject` containing the current array of blocks
- * * A map of ready states
- * from the last save made on that instance
+ * - A map of `EditorJS` instances
+ * - A map of `BehaviorSubject` containing the current array of blocks
+ * - A map of ready states
+ * The are generated from the last save made on that instance
  */
 @Injectable({
   providedIn: 'root'